Place the xml-pubs transformation tools and their tests into their own module, reorganize package structure
